refactor: supabase iac
