+++ plugins/aws/tsconfig.json
@@ -1,5 +1,5 @@
 {
   "extends": "../../tsconfig.base.json",
-  "include": ["src", "sdk", "lambda"],
+  "include": ["src", "sdk", "lambda", "iac"],
   "exclude": ["lib"]
 }
