+++ plugins/supabase/package.json
@@ -14,6 +14,11 @@
     "./edge-functions": {
       "import": "./supabase/index.ts",
       "require": "./supabase/index.ts"
+    },
+    "./iac": {
+      "types": "./dist/iac/index.d.ts",
+      "import": "./dist/iac/index.js",
+      "require": "./dist/iac/index.cjs"
     }
   },
   "license": "MIT",
@@ -34,7 +39,7 @@
     "package.json"
   ],
   "scripts": {
-    "build": "rslib build",
+    "build": "tsup",
     "test:type": "tsc --noEmit",
     "make-migrations": "node --experimental-strip-types ./scripts/make-migrations.ts"
   },
@@ -44,9 +49,11 @@
     "@supabase/supabase-js": "^2.47.10"
   },
   "devDependencies": {
-    "picocolors": "^1.0.0",
     "@hot-updater/postgres": "0.12.7",
     "dayjs": "^1.11.13",
-    "mime": "^4.0.4"
+    "es-toolkit": "^1.32.0",
+    "execa": "^9.5.2",
+    "mime": "^4.0.4",
+    "picocolors": "^1.0.0"
   }
 }
