--- plugins/supabase/rslib.config.ts
@@ -1,14 +1,0 @@
-import { defineConfig } from "@rslib/core";
-
-export default defineConfig({
-  lib: [
-    {
-      format: "esm",
-      dts: true,
-    },
-    {
-      format: "cjs",
-      dts: true,
-    },
-  ],
-});
