+++ plugins/supabase/src/index.ts
@@ -1,3 +1,2 @@
 export * from "./supabaseDatabase";
 export * from "./supabaseStorage";
-export * from "./utils";
