+++ plugins/supabase/tsconfig.json
@@ -1,5 +1,5 @@
 {
   "extends": "../../tsconfig.base.json",
-  "include": ["src"],
+  "include": ["src", "iac"],
   "exclude": ["lib", "supabase"]
 }
