+++ plugins/supabase/tsup.config.ts
@@ -1,0 +1,23 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig([
+  {
+    entry: ["src/index.ts"],
+    format: ["esm", "cjs"],
+    outDir: "dist",
+    dts: true,
+    banner: {
+      js: `import { createRequire } from 'module'; const require = createRequire(import.meta.url);`,
+    },
+  },
+  {
+    entry: ["iac/index.ts"],
+    format: ["esm", "cjs"],
+    dts: true,
+    outDir: "dist/iac",
+    external: ["@hot-updater/supabase"],
+    banner: {
+      js: `import { createRequire } from 'module'; const require = createRequire(import.meta.url);`,
+    },
+  },
+]);
